Add storage and snapshot writer
