+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules/
+ops_board/ops_board.sqlite
+reports/ops-board/
