+++ ops_board/lib/storage.js
@@ -1,0 +1,95 @@
+const path = require("path");
+const fs = require("fs");
+const Database = require("better-sqlite3");
+
+function ensureDir(dirPath) {
+  if (!fs.existsSync(dirPath)) {
+    fs.mkdirSync(dirPath, { recursive: true });
+  }
+}
+
+function defaultDbPath() {
+  return path.resolve(__dirname, "..", "ops_board.sqlite");
+}
+
+function openDb(dbPath) {
+  ensureDir(path.dirname(dbPath));
+  const db = new Database(dbPath);
+  db.pragma("journal_mode = WAL");
+
+  db.exec(`
+    CREATE TABLE IF NOT EXISTS items (
+      id TEXT PRIMARY KEY,
+      source TEXT NOT NULL,
+      source_path TEXT NOT NULL,
+      title TEXT NOT NULL,
+      status TEXT,
+      kind TEXT,
+      raw TEXT,
+      updated_at TEXT NOT NULL
+    );
+  `);
+
+  db.exec(`
+    CREATE TABLE IF NOT EXISTS actions (
+      id TEXT PRIMARY KEY,
+      type TEXT NOT NULL,
+      payload TEXT,
+      created_at TEXT NOT NULL
+    );
+  `);
+
+  return db;
+}
+
+function upsertItems(db, items) {
+  const stmt = db.prepare(`
+    INSERT INTO items (id, source, source_path, title, status, kind, raw, updated_at)
+    VALUES (@id, @source, @source_path, @title, @status, @kind, @raw, @updated_at)
+    ON CONFLICT(id) DO UPDATE SET
+      source=excluded.source,
+      source_path=excluded.source_path,
+      title=excluded.title,
+      status=excluded.status,
+      kind=excluded.kind,
+      raw=excluded.raw,
+      updated_at=excluded.updated_at;
+  `);
+
+  const now = new Date().toISOString();
+  const run = db.transaction((rows) => {
+    rows.forEach((row) => {
+      stmt.run({ ...row, updated_at: now });
+    });
+  });
+  run(items);
+}
+
+function upsertAction(db, action) {
+  const stmt = db.prepare(`
+    INSERT INTO actions (id, type, payload, created_at)
+    VALUES (@id, @type, @payload, @created_at)
+    ON CONFLICT(id) DO UPDATE SET
+      type=excluded.type,
+      payload=excluded.payload,
+      created_at=excluded.created_at;
+  `);
+  stmt.run(action);
+}
+
+function fetchItems(db) {
+  return db.prepare("SELECT * FROM items ORDER BY updated_at DESC").all();
+}
+
+function fetchActions(db) {
+  return db.prepare("SELECT * FROM actions ORDER BY created_at DESC").all();
+}
+
+module.exports = {
+  defaultDbPath,
+  openDb,
+  upsertItems,
+  upsertAction,
+  fetchItems,
+  fetchActions,
+};
