+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "better-sqlite3": "^12.6.2"
+  }
+}
